fixed array copy and introduced RMSprop
