+++ pyml/linear_models/base.py
@@ -26,7 +26,7 @@ class LinearBase(BaseLearner, Predictor):
         self._alpha = alpha
         self._batch_size = batch_size
 
-        if method in ['normal', 'nesterov', 'adagrad', 'adadelta']:
+        if method in ['normal', 'nesterov', 'adagrad', 'adadelta', 'rmsprop']:
             self._method = method
 
         else:
@@ -35,9 +35,10 @@ class LinearBase(BaseLearner, Predictor):
         self._seed = set_seed(seed)
         self._type = _type
 
-        if self._method == 'adagrad' or self._method == 'adadelta' or self._method == 'adadelta':
+        if self._method == 'adagrad' or self._method == 'adadelta' or self._method == 'rmsprop':
             if fudge_factor == 0:
-                warnings.warn("Fudge factor for Adagrad optimisation is 0, it will be set to 10e-8 for your own safety")
+                warnings.warn("Fudge factor for {} optimisation is 0, it will be set to 10e-8 for your own "
+                              "safety".format(self._method))
                 fudge_factor = 10e-8
 
         if self._method == 'adadelta' and self._learning_rate != 1:
