+++ pyml/linear_models/logistic_regression.py
@@ -36,7 +36,8 @@ class LogisticRegression(LinearBase, Classifier):
                         - "nesterov": nesterov method for GD
                         - "adagrad": adagrad method for GD
                         - "adadelta": adadelta method for GD
-        :param fudge_factor: fudge factor for Adagrad and Adadelta to avoid zero divisions
+                        - "rmsprop": rmsprop method for GD
+        :param fudge_factor: fudge factor for Adagrad/Adadelta/RMSprop to avoid zero divisions
 
         Example:
         --------
